Added name returning function

+++ client.h
@@ -10,11 +10,11 @@
     char*   client_get_name(client c);
     int     client_get_num_tax(client c);
     int     client_get_time(client c);
-    int     client_get_bill(client c);
+    float   client_get_bill(client c);
     char    client_get_location(client c);
 
     void    client_set_time(client c, int time);
-    void    client_set_bill(client c, int bill);
+    void    client_set_bill(client c, float bill);
     void    client_set_location(client c, char location);
 
 #endif
